+++ docker-compose.yml
@@ -50,7 +50,6 @@ services:
     environment:
       - MINIO_ACCESS_KEY=minio-access-key
       - MINIO_SECRET_KEY=minio-secret-key
-#      - MINIO_REGION=eu-west-1
     command: minio --compat server data
 
 volumes:
